[libromdata] RpWin32_delayload.h: Added a 'success' variable to DelayLoad_test_##fn().
This is initially set to 0. If DELAYLOAD succeeds, it's set to 1.

This function is thread-safe because the variable is only ever set
to 1 after initial startup. If a race condition occurs, the worst
that will happen is the test function is called twice. The DELAYLOAD
DLL cannot be unloaded once it's loaded.