Reworked the delay load functionality to be a little more usable.
src/delayload/: The header, helper source, and a new .cmake file
are located here. The .cmake file adds /DELAYLOAD entries, since
it's not allowed in #pragma comment(linker). The .hpp file can
be included anywhere, and the .cpp file must be compiled in the
resulting executable. We can't use a static library because MSVC
will just trim it out.

[win32] Compile DelayLoadHelper.cpp.