[libromdata] NCCHReader: Fix CIA decryption.
We need to subtract 16 from phys_addr, not offset. Offset was already
added to phys_addr, so it's only used at this location to determine
if the fixed IV for the first 16 bytes needs to be used.

This fixes a regression from commit a74d9360ee57f5d962132d4b009c244a2e8f8d4e.
([libromdata] NCCHReader: New function readFromROM().)

TODO: CIADecryptor class. This will reduce the complexity of the CIA
decryption code in NCCHReader, and it will make it possible to decrypt
DSiWare CIAs.