[libromdata] SystemRegion: Combined the two internal functions into one.
Both CC and LC are retrieved at the same time.

The public functions both call the same internal function using the same
Once variable, so CC and LC will only be initialized once.

Windows: The Once function always returns TRUE. In the previous commit,
it returned FALSE if the locale name was invalid, but that won't help
any since it will always fail on future attempts.