Initial split of libromdata into librpbase and libromdata.
RomData, RomFields, and all of the classes that aren't system-specific
have been moved into librpbase. The RomData subclasses, RomDataFactory,
and system-specific classes remain in libromdata.

TCreateThumbnail and TImageTypesConfig are still in libromdata.
TCreateThumbnail depends on RomDataFactory, and TImageTypesConfig
depends on the RomData subclasses.

config.libromdata.h.in is now config.librpbase.h.in. The lambda function
check is only needed by libromdata, but I'm keeping it in librpbase
because it doesn't make sense to have a config file for a single thing.

Removed lots of unneeded includes for config.libromdata.h and others,
including <cstdio> in PartitionFile.hpp.

TODO:
- Move everything in librpbase to the LibRpBase namespace.
- Find and remove more unused config.librpbase.h includes.