[libromdata] Nintendo3DSFirm: New parser for Nintendo 3DS FIRM binaries.
I'm not sure if the version number is easily located, so I added a data
file that allows looking up the version number using the CRC32.
This is located in Nintendo3DSFirmData.cpp.

If the ARM11 and ARM9 entry points are both set, this is usually an
official FIRM binary.

If only the ARM9 entry point is set, this is ARM9 homebrew.

TODO:
- Add specific support for getting the version numbers from certain
  well-known ARM9 homebrew, e.g. Luma3DS and GodMode9.
- Boot9Strap has both entry points set, so we need to check for the
  "B9S" string at 0x003D.