[rp-stub] Initial support for multiple desktop environments.
CMakeLists.txt: Get the plugin paths for all supported DEs.
config.rp-stub.h.in: Placeholders for the plugin paths.

rp-stub.c: Added an array with all of the plugin paths. There's also
a 2D array with desktop priorities. This will be used to determine
which plugin to use based on the current DE and what's installed.

TOOD: Determine the current DE. Currently hard-coding XFCE.