[libromdata/tests] GcnFstTest: Ensure numbers are always formatted as en_US.
fstPrint() now takes an optional FstFileCount* parameter. If nullptr, the
file statistics will be printed to the ostream as before. If it isn't
nullptr, the statistics will be copied to the FstFileCount* instead.

NOTE: FstFileCount is the same as fst_stats, which was previously
defined in FstPrint.cpp instead of FstPrint.hpp.

GcnFstTest will take the FstFileCount and manually print the numbers
using print_uint32_nolocale(), which is a simplistic function that
prints a uint32_t using en_US formatting.

This fixes issue #70. (tests fail on local different than en_US)