Use CMake's CONFIGURE_FILE() to automatically update manifest versions.
The manifest files are now .in files. They have replacement tags for
the `version` and `processorArchitecture` fields, which are filled in
by CMake's CONFIGURE_FILE() at configure time.

This will ensure that I don't forget to update the manifest version
numbers before release (which I did for rpcli in v1.0).

[cmake] win32.cmake: Set WIN32_MANIFEST_PROCESSOR_ARCHITECTURE to the
correct value for the manifest's `processorArchitecture` field.