[kde] ImageTypesTab: Added a QSignalMapper for the QComboBoxes.
Changing a QComboBox now results in the "Apply" button being enabled.
Clicking "Reset" results in the QComboBoxes being reset and the
"Apply" button being disabled.

cboImageType_priorityValueChanged(): This function now returns a bool
indicating if the value was actually changed. This is used to determine
if the modified() signal should be emitted.