[libromdata] TImageTypesConfig::reset(): Don't unconditionally reset all ComboBoxes to 0xFF.
This causes some flickering on Windows when clicking "Reset".

Instead, maintain an array of bools indicating which ComboBoxes need to
be reset. These are all `true` initially, and they're set to `false`
whenever a valid priority is found for a ComboBox.

Any remaining ComboBoxes with the value set to `true` is reset to "No".