[kde] DownloadsTab: Implemented the Downloads tab.
reset(), loadDefaults(), and save() are all implemented and working.

QSettings saves boolean values as "true" and "false", so no hackish
workarounds are needed for the Downloads section.

ConfigDialog.ui: Added the Downloads tab as well as some missing
slot definitions for ImageTypesTab (though they aren't going to
be used in the UI, at least right now).