[libromdata] CIAReader: New IDiscReader implementation for decrypting CIA contents.
This reader can only be used to read individual CIA contents, not the
full CIA. The CIA header must be parsed and handled separately.

NCCHReader: Removed all CIA decryption functionality. To decrypt a
CIA, create a CIAReader and pass it to the constructor.
- NOTE: NCCHReader *takes ownership* of IDiscReader subclasses.
  This may be changed if IDiscReader is rewritten to be a subclass
  of IRpFile, but this is how it works for now. I did this mostly
  to make it easier to create a temporary CIAReader without having
  to keep track of it. (Maybe use refcounting instead?)

TODO: Use CIAReader for DSiWare. May need to add IDiscReader support
to the NintendoDS class (or make IDiscReader a subclass of IRpFile).