[kde] ITab: Added a "defaults" property.
Tabs can indicate whether or not they want to have the "Defaults" button
enabled. This is useful for tabs that don't have default values.

KeyManagerTab: Return false for hasDefaults(), since the "default" value
for keys would be all empty keys, and that's not very useful.

ConfigDialog: Enable/disable the "Defaults" button in response to
switching tabs.