[kde] KeyManagerTab: Added a dropdown menu to the "Import" button.
The menu contains the three import actions. Note that none of the actions
are actually implemented yet, so the menu doesn't do anything.

KeyManagerTab.ui: Added a mnemonic to "Import" ("I&mport"). KDE was
adding this automatically, but I decided to make it explicit. 'm' is
used instead of 'I' because 'I' is used by the "Image Types" tab.