[kde] KeyManagerTab: Implemented the modified() signal.
KeyStore: Emit modified() whenever the user changes a key.
This is different from the keyChanged() signals, which are emitted
even when resetting the keys to their original values.

KeyStore::modified() is forwarded to KeyManagerTab::modified().

Moved `changed` from KeyManagerTab to KeyStore. It's currently
cleared on reset() and set whenever a key is modified by the user.

ConfigDialog: Connected the modified() signal from KeyManagerTab.
Call KeyManagerTab's save() and reset() functions when needed.

KeyStore::setKey(): Compare the value *after* converting to uppercase
and/or from kanji to hexadecimal.