[kde] KeyStore: New function binToHexStr().
This converts a binary key to a hexadecimal QString. It's now used in
various areas that used hex_lookup[] directly before.

It's NOT used in the kanji code, since that needs to convert UTF-16
code points to hexadecimal, not just bytes.

Use `pKey` to refer to pointers to Key objects instead of `key`.