[kde] KeyStore: Decrypt encrypted keys in aeskeydb.bin if Slot0x2CKeyX is available.
getAesKeyDB_key(): New function that gets the actual key used for the
aeskeydb.bin encryption. This uses Slot0x2CKeyX and a NULL KeyY,
scrambled with the CTR key scrambler.

The counter value is the first 16 bytes of the AesKeyInfo struct.

TODO: Support aeskeydb.bin files encrypted with the debug key?