[xfce] rp-thumbnail-dbus: Converted from dbus-glib to GDBus.
Mostly based on xfce's tumbler:
https://github.com/xfce-mirror/tumbler

Instead of exporting the RpThumbnail object directly, a skeleton object
created using gdbus-codegen is exported. The server-side object has
signals that map to the D-Bus methods.

RpThumbnail no longer has signals that map directly to the D-Bus object.
Instead, it calls functions in the skeleton object, which export the
signals.

Unfortunately, the file size has increased a bit. (TODO measure file
sizes in debug and release.)

FindGLib2.cmake: We need gio-unix for the gdbus-codegen generated files.
Import it in GLib2::glib so we don't have to link to it manually in the
rp-thumbnail-dbus project. (TODO: Make this a GIO library target instead
when doing the main commit)

TODO:
- Split the RpThumbnail object from the main code.
- Move rp-thumbnail-dbus out of the xfce directory?