[libromdata] NCCHReader::openLogo(): Support SDK5+ logo regions.
The SDK5+ logo region is plaintext.

n3ds_structs.h: Added a pseudo-section N3DS_NCCH_SECTION_PLAIN.
NCCHReader handles this section by not decrypting anything.

Interestingly, this means that we can check the logo in encrypted
SDK5+ .3DS files if we don't have the keys, but we can't check it
for pre-SDK5 .3DS files.