[gtk] CreateThumbnail: Save XDG cache keys in the thumbnail file.
These keys are used to detemrine if the thumbnail should be re-cached.

tumblerd no longer forces thumbnail regeneration when Thunar's view is
refreshed.

GNOME uses these keys as well, but it managed the keys itself, so the
thumbnailer didn't need to add them. (GNOME seems to remove the keys
we're adding now, too.)

KDE also uses these keys, but the KDE plugins don't write to the thumbnail
cache directly. We return a QImage, and KDE's thumbnail.so handles it.