[librpbase] RpPngWriter: New function write_tEXt().
This function allows writing a vector of key/value pairs using tEXt
chunks. I was originally going to use this to handle the XDG
thumbnailing fields, but it turns out I have to use GdkPixbuf and/or
QImage because of the image scaling functionality.

I'll leave this here anyway, since it could be useful later.