[kde] KeyStoreModel: Reworked to use a tree model. (QAbstractItemModel)
Note that instead of keeping a separate copy of all of the nodes, the
KeyStore is used directly. QModelIndex's internal ID is used to keep
track of the parent/child relationships. (LOWORD is the section index;
HIWORD is the key index, or 0xFFFF for the section header.)

Enabled root decorations so the arrow can be seen.

Section headers have the First Column Spanned property set, so they
aren't constrained to the first column.

KeyManagerTab: Delete keyStoreModel before keyStore. There seems to
be some race conditions with the way I have things set up that result
in the KeyStore object being used after it's deleted if we don't
delete the keyStoreModel first.