[libromdata] KeyStoreUI: Emit modified_int() when importing keys.
FIXME: Importing Wii keys.bin results in "1 key not imported because
it's invalid", since it tries importing both Retail and Debug using
the same slot. It should only report the correct key.

Also, the problem with aeskeydb.bin on Windows was because I didn't
have ctr-scrambler set. It works fine if ctr-scrambler is set.