[win32] KeyManagerTab: Implemented ListView editing.
The ListView control only natively supports editing the first column,
so an EDIT control overlay is used. This requires subclassing both
the ListView and the EDIT control to provide full functionality and
to intercept dialog keys so they don't close the dialog.

Note that ListView_SetItem() doesn't work with LVS_OWNERDATA, so the
KeyStore is modified directly and a redraw is requested.

The EDIT control is created as a child control of the ListView. This is
needed in order to prevent drawing/focus issues.

The ES_UPPERCASE style is used to force hexadecimal digits to uppercase.

References:
- http://www.cplusplus.com/forum/windows/107679/
- https://support.microsoft.com/en-us/help/102589/how-to-use-the-enter-key-from-edit-controls-in-a-dialog-box