[librpbase] TextFuncs: New function rp_sprintf().
This is an sprintf()-like function that takes UTF-8 strings, calls
vsnprintf(), and returns the value as an rp_string.

In RP_UTF16 configurations, the UTF-8 result is automatically
converted to UTF-16.

Note that this may have some extra overhead due to a memory allocation,
but it helps to reduce the number of lines required to properly format
string data everywhere else.

Other changes in this commit:
- EXE: If `machine` was an unknown value, and the executable was marked
  as .NET, the string " (.NET)" would be added to CPU twice. This is
  a very unlikely scenario, though...
- DMG: Cast `rom_size` to `unsigned int`. It's only negative if an error
  occurred, and that's checked before calling rp_sprintf().
- Some other `unsigned int` changes elsewhere in DMG and in NES.