[win32] KeyManagerTab: Added an icon for "good" keys. (hIconGood)
Icon is based on KDE Oxygen 5.35.0's base/16x16/actions/dialog-ok-apply.png

The .ico file has 32-bit, 8-bit, and 4-bit variants.

Note that I used an icon instead of a bitmap (or PNG converted to bitmap)
because AlphaBlend() requires premultipled alpha, and PNG doesn't use
premultiplied.

This is currently the only icon resource in the DLL, so it might be
misinterpreted as an "application" icon. Maybe I should add more icons...