[win32] RP_ShellPropSheetExt: Use alternating row colors in RFT_LISTDATA.
SubtabDlgProc(): Forward NM_CUSTOMDRAW messages to the parent dialog.
This is needed because if tabs are used, the ListView controls are
technically owned by the tab child dialogs, not the main dialog.

FIXME: There seems to be some extra uncolored space on the left side
of the first column. This appears on KeyManagerTab, too.