[win32] KeyManagerTab: Draw the background for "Valid?" even for main rows.
This works around an issue where ListView only partially recognizes
theme changes. It notices theme changes for items not totally drawn
by the CUSTOMDRAW handler, but not for group headers or padding.

This also fixes a glitch on XP where resizing the "Valid?" column would
result in ghosting of the icon in that column. I also noticed that XP
seems to clip the icon on the column boundaries, whereas 7 does not.