[libromdata] Nintendo3DS: Show the service listing.
It's unlikely that a title wouldn't use any services, because then it
wouldn't be able to do much. However, if that happens, it simply won't
display the "Services" field.

Reduced `rows_visible` on the other permissions RFT_LISTDATA fields
from 8 to 6 in order to fit everything on Windows.

TODO: ext_services? (9.3.0-X)