[gtk] RomDataView: Added support for RFT_LISTDATA rows_visible.
The initial resize is done when the GtkTreeView emits the "realize"
signal. This is needed because the sizes aren't known prior to
realization.

It took quite a while to get everything working for both GTK2 and
GTK3. Among other things, the GtkTreeView header is composed of
one widget per column, and in GTK2, there's no accessor function,
so we have to break through GSEAL if it's enabled.

Also, on Kubuntu 17.04 (gtk-2.24.30-4ubuntu3), GTK2 reported the header
widget as having a height of 0px if it wasn't visible, whereas on Gentoo
(gtk-2.24.31-r1) its normal size was reported. Hence, we need to check
the GtkTreeView's header visibility property instead.

TODO:
- Handle font/style changes.
- Check Qt4 for realize? I didn't need that for Qt5...]