[librpbase] RomFields::ageRatingsDecode(): Moved the age ratings string building code here.
It was duplicated in the GTK+, KDE, and Win32 UIs, as well as rpcli.
There might be some slight overhead with the new one on KDE due to a
string copy, but the code reduction outweighs that overhead.

Ironically, the original age ratings field was a string field. I changed
it to the age_ratings_t array to make it more usable for things other than
displaying text, but I haven't used it for anything other than text yet.
Maybe I'll have it display logos later...

[rpcli] Use RomFields::ageRatingsDecode() in the plain text output.
This version uses the `newlines` parameter to disable automatic newlines
after every four ratings. The JSON output still uses custom code due to
JSON formatting.