[win32] RP_ExtractImage: Use CreateThumbnailNoAlpha.
I was originally going to add a template parameter to CreateThumbnail,
but it seems that there's no easy way to use the class template parameter
in a function without specializing it, i.e. it can't be passed as a
variable. I'd have to make duplicate copies of each function, which
defeats the point of the split.

Instead, create a new class, CreateThumbnailNoAlpha, that inherits from
CreateThumbnail and replaces rpImageToImgClass() and rscaleImgClass()
with its own code.

Alternatively, I could add two new virtual functions to CreateThumbnail
for the image conversion and have CreateThumbnail's functions call those,
then override the new virtual functions in CreateThumbnailNoAlpha, but
that's probably just more complexity for the sake of complexity.