[libromdata] DirectDrawSurface: Added support for linear ARGB1555 and RGB565 images.
Moved the RGBA and RGB format tables out of loadFieldData() and into the
private class. Added accessor functions to retrieve the description field
and the new px_format field (stored as uint8_t to save memory).

NOTE: We can't store the description as an rp_char array due to issues
with _RP() on MSVC versions older than 2015.

[librpbase] ImageDecoder::fromLinear16(): Added a `pitch` parameter.
This is required for DirectDrawSurface because the texture might have
a pitch that's larger than the texture width (multiplied by bytespp).