[librpbase] ImageDecoder::fromLinearRGB(): Untemplated the `px_format` parameter.
With the number of pixel formats I'm going to add for DDS, the template
expansion is going to grow enormously. The format selection code is a
single `switch` segment, so the runtime overhead is minimal compared to
the template expansion overhead.

Note that I'm going to keep uint16_t, uint32_t, and 24-bit as separate
functions instead of using a template parameter for those.

[libromdata] DirectDrawSurfacePrivate::loadImage(): Changed the main
`switch` statement to switch based on `bytespp` instead of `px_format`.