[libromdata] DirectDrawSurface: Change the table split from alpha to bpp.
This should reduce the amount of comparisons needed to find the correct
format, since there's fewer entries for bpp checks than for alpha vs.
no alpha checks.

Updated the search functions to set a pointer to the start of the table
that matches the bpp. The table is then searched until an entry with a
NULL description is encountered. This takes up fewer lines of code
compared to using ARRAY_SIZE() for a counter.

There is now only one table struct, RGB_Format_Table_t, and it contains
the `Amask` field. Note that `bits` is no longer present, since it's
implied by which table is in use.