[librpbase] ImageDecoder_S3TC.cpp: Fix tile assertions for DXT1, DXT3, and DXT5.
GameCube DXT1 requires width and height to be a multiple of 8 due to its
use of 2x2 blocks of 4x4 tiles. Standard DXT1, DXT3, and DXT5 only has
the 4x4 tiles, so those should check for multiples of 4, not 8.

This fixes the display of SA2 PC's tento_alert*.dds textures.