[librpbase/tests] Added an ImageDecoder test.
Currently only handles DDS textures. The only test cases at the moment
are for DXT1, DXT3, and DXT5 with S3TC enabled.

The test images are an RGBK square, with an alpha variant that increases
alpha in the lower-right quadrant. The code used for generating this
test image can be found in ref-gen.cpp.

rp_image: New function dup_ARGB32(). This function is similar to dup(),
but it converts the image to ARGB32 if it's currently in CI8. This is
used to allow for easier comparisons in ImageDecoderTest.