[librpbase] ImageDecoder::fromDXT1_GCN(): Eliminated one of the 2x2 tiling loops.
Do all four tiles at once instead of separate loops for `x` and `y`, then
blit all four tiles at once.

This seems to increase the code size a bit in the Linux debug build due
to the extra BlitTile<>() calls, but the loop reduction should make up
for that.