[librpbase] AesCAPI, AesCAPI_NG: Use 64-bit XORs in the CTR path.
This is automatically reduced to 32-bit XORs in 32-bit builds.
It should be faster overall than manually XORing each individual
byte, plus we eliminated the loop.

Alignment shouldn't be an issue, since this is Windows-specific code and
I'm only targetting i386/amd64. It might be an issue on ARM, but Windows
on ARM is a wasteland.

NOTE: MSVC 2017 seems to be optimizing this to SSE2 code, using `movups`
for 32-bit and `movaps` for 64-bit. gcc-7.1.0 also uses SSE2 code in
32-bit builds if `-msse2` is specified, but uses `movq` for 64-bit chunks
instead of doing the whole thing at once. gcc doesn't use SSE2 in 64-bit
builds, though.