[librpbase] RpJpeg: Use libjpeg-turbo's JCS_EXT_BGRA if available.
libjpeg-turbo supports converting to 32-bit RGBA, so we should use its
conversion instead of our own, since it's probably more optimized.

Note that JPEG calls it "BGRA" whereas we call it "RGBA".

Since there's no way to determine if JCS_EXT_BGRA is available without
trying it, I expanded the setjmp() code to use two levels. If it fails
the first time when it tried JCS_EXT_BGRA, it will try another time
without it.

Use MY_JCS_EXT_BGRA instead of #defining JCS_EXT_BGRA, since we might
be compiling with the original libjpeg instead of libjpeg-turbo, or
we might be compiling with libjpeg-turbo.

Since JCS_EXT_BGRA matches ARGB32, if it's used, we set `direct_copy`
so we end up loading the scanlines directly into the rp_image.

Added an output_message() override. libjpeg's default output_message()
uses MessageBox(GetActiveWindow(), ...) on Windows, which is really
bad and user-hostile. Our output_message() uses OutputDebugStringA(),
which allows you to see the message if a debugger is connected but
otherwise is handle as a no-op. On Linux, fprintf(stderr, ...) is used.