[libromdata] Dreamcast: Stop parsing the directory table once we encounter an entry with a 0 length.
The directory might not take up the full amount of space reserved due
to blocks being 2048 bytes long, so we should exit once we hit a 0-length
record instead of infinitely looping.

This fixes the lockup with Atari Anniversary Edition and ChuChu Rocket!.

Atari Anniversary Edition doesn't seem to have a 0GDTEX.PVR file, but
ChuChu Rocket! does. It's not loading for some reason, though...