[win32] RP_ShellPropSheetExt: Don't keep the file open for long periods of time.
We should open it initially to detect the file type, then close it, then
reopen it when the "ROM Properties" tab is selected.

After the "ROM Properties" tab is selected, the RomData object is kept
around, but the underlying file is closed.

We're not keeping the original file open because we'd have to load the
field data and images at that point, and if the user doesn't click the
"ROM Properties" tab, that's a waste of time.

This fixes an issue where the user can't enable NTFS compression on a
file that's handled by ROM Properties.