[libromdata] Nintendo3DSSysTitles: New database with several 3DS system titles.
These are mostly titles that would be found in official badges.
Note that we're not using fancy tricks to determine the region, since
that's not necessarily accurate for many titles outside of 0x00040010.
Instead, all 6 title IDs are listed.

TODO: Use bsearch()?

NintendoBadge: Use Nintendo3DSSysTitles to decode the title ID.