[libromdata] TCreateThumbnail: We need the ImgClass's size, not the rp_image's size.
On Windows, the ImgClass is squared because Windows doesn't really like
non-square icons or thumbnails. However, we were returning the original
image width, which caused a problem with the nearest-neighbor rescaling
function. This only showed up with the addition of NintendoBadge, since
the only other class that had non-square images was Amiibo, and that
class doesn't use nearest-neighbor rescaling.

This fixes issue #93. (improper handling for several badges)
Reported by @cheatfreak.