[librpbase] RpPngWriter: Write the rp_image's sBIT metadata to the PNG file.
rp_image::sBIT has the same format as png_color_8, so we can use it
directly with reinterpret_cast<>.

FIXME: libpng is complaining if sBIT has alpha == 0, since we're telling
it that the image is PNG_COLOR_TYPE_RGB_ALPHA. We need to use the color
type PNG_COLOR_TYPE_RGB if alpha == 0. This will also reduce the size
of the PNG image, since it won't store an alpha channel.

Note that this optimization only works if the code that creates the
rp_image objects sets sBIT correctly. If sBIT isn't set at all, it will
act like it did before. (No sBIT chunk, always alpha.)