[librpbase] RpPngWriter: Set the PNG color type to PNG_COLOR_TYPE_RGB if `sBIT.alpha == 0`.
This fixes writing the sBIT chunk, and it reduces the size of PNG images
that don't have alpha channels.

write_CI8_palette(): Always write a tRNS chunk, even if `sBIT.alpha == 0`,
since sBIT's alpha value doesn't affect paletted images.

write_IDAT(): If skipping alpha, and the source image is ARGB32, add a
filler byte. This tells libpng to skip the alpha byte from the source,
so we don't have to manually remove it.