[librpbase] RpPngWriter: Removed the `const char *filename` private class constructors.
The public class constructors can create an RpFile and pass that RpFile
directly to the private class constructors.

The public class constructors that take IRpFile now dup() it and pass
the dup()'d version to the private class constructors.

TODO: Unlink the file if an error occurs. Not sure if I can easily
implement this in RpPngWriter, or if I should do this in the caller.