[librpbase] RpPngWriter: Use PNG_Z_DEFAULT_COMPRESSION.
PNG_Z_DEFAULT_COMPRESSION == -1, but it seems to correlate to level 6,
which is slightly better without being that much slower.

Four Swords Anniversary's icon was reduced from 3,201 to 2,193 bytes
when using rp-stub to thumbnail with rom-properties-gnome.so with
RpPngWriter after this change.

TODO: Make it customizable, and maybe use filters?