[librpbase] RpPngWriter: Use UTF-8 encoding for values and compress text.
If the string can be fully represented as ASCII, it's stored as-is.
If it can be fully represented as Latin-1, it's converted from UTF-8
to Latin-1 and then stored. Otherwise, it's stored using iTXt.

If the string is >= 40 bytes, it will be stored as zTXt. This is the
same threshold as Qt.

FIXME: libpng-1.4 fully supports iTXt, but older versions might not.
Need to test this on Ubuntu. Note that Ubuntu has libpng-1.6 in the
repository, but the default version is libpng-1.2.

TODO: Move the Latin-1 code to TextFuncs. Maybe use system encoding
functions instead of our own.