[librpbase] RpGdiplusBackend: Ensure `stride` is a multiple of 16 bytes.
This will allow us to use more optimal SSE2 code.

The only way to specify `stride` is to manually allocate a buffer when
locking the GDI+ bitmap, so we do that now. m_pImgBuf is allocated on
the first lock() and freed when the backend is destroyed.

Updated the bitmap conversion functions to use the correct stride for
HBITMAP. HBITMAP stride is always a multiple of 4, so no adjustment is
necessary for ARGB32, but alignment is needed for CI8.