[librpbase] FileSystem.hpp: New function resolve_symlink().
Moved the Linux-specific code form RelatedFile.cpp to FileSystem_posix.cpp
as resolve_symlink().

Implemented resolve_symlink() on Windows using GetFinalPathNameByHandleW().
Note that this function was added in Windows Vista, which is also the
first version of Windows to support symlinks in general. Hence, we have
to get the function pointer using GetProcAddress().