[libromdata] GdiReader: New function openTrack().
This opens the specified track number (1-based) if it isn't opened yet.
Track 03 and the last data track (if not 03) are now automatically opened
by the private class constructor. If this fails, the tracks will be closed
and the disc reader will not be usable.

Set block_size and calculate blockCount and disc_size based on the
last data track. A full Dreamcast disc has 549,150 sectors.