[libromdata] SegaPVR: Handle GBIX lengths other than 8 (16).
The GBIX struct has a length field. Normally PVRs with GBIX have this
field set to 8, which means the entire GBIX struct is 16 bytes. However,
Shenmue's 0GDTEX.PVR files have a GBIX with length == 4, which makes the
GBIX struct 12 bytes.

We need to handle this in order to get Shenmue's 0GDTEX.PVR image to
show up correctly.

Sanity check: Length must be in the range [4,128].

Thanks to @SonicFreak94 for pointing this out.