[kde] ImageTypesTab: Split the headers and combo boxes into separate grids.
This will make it easier to add a scroll bar that only scrolls the image
types and not the headers.

ColumnResizer: Column synchronization for Qt. License: BSD 3-clause
References:
- http://agateau.com/2011/clean-up-your-layouts-with-columnresizer/
- https://github.com/agateau/columnresizer

Some modifications were necessary to get it to compile with Qt5.

NOTE: The combo boxes are slightly wider now for some reason.
Not sure if I can easily shrink them.