[win32] AboutTab: Show the icon.
This is a 128x128 icon. It's currently the same icon used for the window.

PropSheetIcon: Get a 128x128 variant of the icon. On Windows 7, a 256x256
version is available, so that's downscaled. On Windows XP, 48x48 is the
largest version, so that's upscaled.

Reposition the labels and tab control after loading the icon.

Repositioned the labels such that they're fully center-aligned if the
icon could not be loaded.