[librpbase] ImageDecoder_Linear_sse2.cpp: Fixed and slightly optimized ARGB1555/RGBA5551 alpha handling.
The new version for ARGB1555 uses 8-bit comparisons with corresponding
masks in order to eliminate the extra low-byte mask. Note that SSE2
comparisons are *signed*, so in order to match 0x8000, we have to compare
the high byte as "less than 0x00", and the low byte as "less than 0x80".

For RGBA5551, we use an 8-bit comparison and a left shift in order to
fix the masks. This eliminates a memory access.

This also fixes the following tests:
[  FAILED  ] 3 tests, listed below:
[  FAILED  ] fromLinear16/ImageDecoderLinearTest.fromLinear_cpp_test/PXF_ARGB1555_0xFF218CA5, where GetParam() = PXF_ARGB1555_0xFF218CA5
[  FAILED  ] fromLinear16/ImageDecoderLinearTest.fromLinear_sse2_test/PXF_ARGB1555_0xFF218CA5, where GetParam() = PXF_ARGB1555_0xFF218CA5
[  FAILED  ] fromLinear16/ImageDecoderLinearTest.fromLinear_sse2_test/PXF_ABGR1555_0xFF218CA5, where GetParam() = PXF_ABGR1555_0xFF218CA5

[libromdata/tests] ImageDecoderLinearTest: Fix alpha for ARGB1555.