[librpbase] Implemented IFUNC-based dispatch for ImageDecoder.
This uses the ELF "IFUNC" symbol type to implement a resolver function.
The resolver function returns a function pointer to either the C++,
SSE2, or SSSE3 versions of the functions, if available.

This allows us to reduce dispatch overhead to only the first invocation.

cpu_dispatch.h: New file with CPU and IFUNC detection macros.
Use IFUNC_INLINE to allow the dispatch function to be inlined if IFUNC
isn't supported.

config.librpbase.h.in: Check for `features.h`, which is required in
order to determine the glibc version.

ImageDecoder_ifunc.cpp: New file with IFUNC resolver functions for
ImageDecoder. Note that the resolver functions have to be `extern "C"`.

CMakeLists.txt: Only compile ImageDecoder_ifunc.cpp if we're on i386
or amd64 (since those are the only ones with optimizations) and if
we're building for non-Apple Unix systems, since only glibc is currently
supported. Note that this *will* end up building on other Unix systems
that don't have glibc, but the preprocessor checks will end up NOPing
out the entire file.

TODO: Check for IFUNC on other operating systems, e.g. the BSDs.