[kde] RpQImageBackend: Allocate our own aligned image buffer.
Qt5 has an optional QImageCleanupFunction parameter, which lets it
delete the image buffer when it's no longer needed. This effectively
gives Qt5 ownership of the buffer.

Qt4 does *not* have this parameter, so we have to free the buffer
ourselves. This could be a problem if a QImage is taken and still
points to the rp_image data after the rp_image is deleted.

FIXME: Need to test on a KDE4 system to make sure this still works.