[gtk] GdkImageConv_ssse3.cpp: Make sure the temporary palette is 16-byte aligned.
NOTE: rp_image_backend_default guarantees 16-byte alignment for the
palette data, as does RpGdiplusBackend. RpQImageBackend, however,
uses QVector for the palette, since that's required for QImage,
so it does *not* have 16-byte alignment. Hence, we can't easily add
general SSE2/SSSE3 palette conversion optimizations right now.

For GTK+, it should be fine.

NOTE 2: We could probably use an attribute for alignment on the stack,
but we're better off allocating the palette on the heap.