[librpbase] RpGdiplusBackend: Correctly align the palette.
Gdiplus::ColorPalette has two UINT entries before the color table.
This means we have to offset the aligned data by 8 bytes in order
to get the color table to have 16-byte alignment.

Note that nothing currently depends on the palette being aligned
at the moment, but it might later on.