[librpbase] CMakeLists.txt: Use CHECK_SYMBOL_EXISTS() instead of CHECK_FUNCTION_EXISTS() for aligned malloc().
CHECK_FUNCTION_EXISTS() gets confused if we have `/Gz` specified for
stdcall because it doesn't include any headers, so MSVC will assume the
function should be stdcall, which is wrong. Hence, it won't compile
because it won't find the correct symbol.

CHECK_SYMBOL_EXISTS() allows us to specify the header (or headers),
so MSVC will know the correct calling convention.

Added a CMake check to make sure that at least one aligned malloc()
function was found. This is better than relying on the check in
`aligned_malloc.h`, since that happens later in the build.