[win32] RP_ShellPropSheetExt: Optimize UTF-16 text conversion.
When we have to both measure and display text, convert it to UTF-16 once
and store the result instead of converting to UTF-16 twice. This should
improve performance a bit at the cost of memory overhead, though it will
also reduce memory allocations due to how std::wstring works.

Use the localized "%s:" to measure the field description label width.