[libromdata] Combined the GL enumerations into a single array.
It seems that not all of the fields in the KTX reference files are
limited to the sections mentioned in the KTX structure reference,
so let it use everything.

There's now a single array and single lookup function. It shouldn't
be too much slower than the individual functions due to how binary
searches work, i.e. it's O(log(n)).

Renamed GLTypeStrings to GLenumStrings.