[librpbase] ImageDecoder::fromETC2_RGB_A1(): Added untested support for ETC2 with punchthrough alpha.
Unlike ETC2 RGB or RGBA, this eliminates the "individual" mode so it can
repurpose the diffbit as an opacity bit.

For differential mode only, a separate intensity table is used if
opaque == 0 and the pixel index is not 2.

For differential, 'T', and 'H' modes, if opaque == 0 and the pixel
index is 2, the pixel is RGBA(0,0,0,0). Otherwise, it's handled like
regular ETC2.

Note that this does not apply for 'Planar' mode, which is always opaque.

KhronosKTX: Added support for GL_COMPRESSED_RGB8_PUNCHTHROUGH_ALPHA1_ETC2.