[libromdata/tests] Regenerated RGB/RGB565.dds.gz.
For some reason, this was stored as DXT1 instead of RGB565.
The test passed because the reference image was based on this image
instead of being generated separately.

This fixes the VTF BGR565 and RGB565 tests, since the reference image
now matches the actual image contents.

TODO: Verify that the image encoding matches the test.