[libromdata] ValveVTF3: Preliminary support for Valve VTF3 (PS3) format.
This format is used in Valve's PS3 games, including Portal.

It's completely undocumented as far as I know, so this will take more
effort than regular VTF. I have width and height identified, and it
seems that the actual texture data is usually DXTn. Figuring out which
DXTn programmatically will be more difficult, as will figuring out if
there's any mipmaps that need to be skipped.