[libromdata] ValveVTF3: Decode all images as DXT1.
Since we don't have a proper mipmap count yet, we'll just assume that
the texture is stored at the end of the file.

TODO: A lot of the textures are DXT5 and thus show up wacky if decoded
as DXT1, since it's trying to decode alpha as if it's RGB. Figure out
if there's a way to distinguish DXT1 vs. DXT5.