[libromdata] ValveVTF3: Found the DXT1 vs. DXT5 flag.
vtf3Header.flags, bit 13 (0x2000), indicates whether or not an alpha
channel is present. If set, there is an alpha channel, so the image is
encoded using DXT5. If not set, there's no alpha channel, so the image
is encoded using DXT1.

TODO: Some images have bit 7 (0x0080) set, but I'm not sure what that
bit means...