[libromdata] SNES: Initial support for BS-X ROM images.
The BS-X ROM header is similar to SNES ROM headers, though stuff is
reorganized a bit, so I created a union of structs.

There's no easy way to identify a BS-X ROM compared to an SNES ROM,
with the exception of certain memory pack ROMs. This detection isn't
supported, but I left a comment in the SNES::SNES() constructor.

NOTE: While BS-X ROMs are detected now, the field data that's displayed
is incorrect, since it's using the SNES header offsets.

Thanks to LuigiBlood for providing sample ROM images.

References:
- http://problemkaputt.de/fullsnes.htm
- http://satellaview.wikia.com/wiki/Satellaview_ROM_header